sdk: move `LimitedData` to sdk-common
