+++ benchmarks/src/main/scala/org/typelevel/otel4s/benchmarks/MultiSpanProcessorBenchmark.scala
@@ -105,7 +105,8 @@ object MultiSpanProcessorBenchmark {
       import org.typelevel.otel4s.trace.{SpanContext, SpanKind}
       import org.typelevel.otel4s.sdk.TelemetryResource
       import org.typelevel.otel4s.sdk.common.InstrumentationScope
-      import org.typelevel.otel4s.sdk.trace.data.{LimitedData, SpanData, StatusData}
+      import org.typelevel.otel4s.sdk.data.LimitedData
+      import org.typelevel.otel4s.sdk.trace.data.{SpanData, StatusData}
       import org.typelevel.otel4s.sdk.trace.processor.SpanProcessor
 
       val processor: SpanProcessor[IO] = new SpanProcessor[IO] {
@@ -124,8 +125,8 @@ object MultiSpanProcessorBenchmark {
         endTimestamp = None,
         status = StatusData.Ok,
         attributes = LimitedData.attributes(Int.MaxValue, 1024),
-        events = LimitedData.events(Int.MaxValue),
-        links = LimitedData.links(Int.MaxValue),
+        events = LimitedData.vector(Int.MaxValue),
+        links = LimitedData.vector(Int.MaxValue),
         instrumentationScope = InstrumentationScope.empty,
         resource = TelemetryResource.empty
       )
