+++ sdk/trace/src/main/scala/org/typelevel/otel4s/sdk/trace/data/LinkData.scala
@@ -22,6 +22,7 @@ package data
 import cats.Hash
 import cats.Show
 import cats.syntax.show._
+import org.typelevel.otel4s.sdk.data.LimitedData
 import org.typelevel.otel4s.trace.SpanContext
 
 /** Data representation of a link.
